Implement FormatMountLabel unconditionally
Implementing FormatMountLabel on situations built without selinux
should be possible; the `context` will be ignored if no SELinux is
available.
